requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.0
django-filter>=23.1
psycopg2-binary>=2.9  # only needed for PostgreSQL
